requirements.txt
python-telegram-bot==21.9
httpx==0.27.2
aiohttp==3.9.5
tzdata

